Refactor and fix tests
