+++ isimip_utils/files.py
@@ -8,7 +8,7 @@ logger = logging.getLogger(__name__)
 
 
 def find_files(base_path: str | Path, file_iter: Iterable[Path],
-               pattern: str = r'_(?P<start_year>\d{4})_*(?P<end_year>\d{4})?\.nc\d*$') -> tuple[list[tuple], int, int]:
+               pattern: str = r'_(?P<start_year>\d{4})_(?P<end_year>\d{4})?\.nc\d*$') -> tuple[list[tuple], int, int]:
     """Find files for a given (dataset) path, matching a regex pattern for start and end year.
 
     Args:
