+++ isimip_utils/tests/test_utils.py
@@ -4,13 +4,8 @@ from isimip_utils.exceptions import ValidationError
 from isimip_utils.utils import (
     Singleton,
     cached_property,
-    copy_placeholders,
     exclude_path,
-    get_permutations,
-    get_placeholders,
     include_path,
-    join_parameters,
-    update_year,
     validate_lat,
     validate_lon,
 )
@@ -20,11 +15,6 @@ paths = [
     'a/b/d',
     'a/b/e'
 ]
-
-parameters = {
-    'model': ['model_a', 'model_b'],
-    'variable': ['x', 'y', 'z']
-}
 
 
 def test_singleton():
@@ -87,62 +77,3 @@ def test_include_path():
     assert include_path(paths, 'a/b/c') is True
     assert include_path(paths, 'a/b/cc') is True
     assert include_path(paths, 'a/b/f') is False
-
-
-def test_get_permutations():
-    assert get_permutations(parameters) == (
-        ('model_a', 'x'),
-        ('model_a', 'y'),
-        ('model_a', 'z'),
-        ('model_b', 'x'),
-        ('model_b', 'y'),
-        ('model_b', 'z')
-    )
-
-
-def test_get_placeholders():
-    assert get_placeholders(parameters, ('model_a', 'x')) == {
-        'model': 'model_a',
-        'variable': 'x'
-    }
-
-
-def test_join_parameters():
-    assert join_parameters(parameters) == {
-        'model': 'model_a+model_b',
-        'variable': 'x+y+z'
-    }
-
-
-def test_join_parameters_max_count():
-    assert join_parameters(parameters, 2) == {
-        'model': 'model_a+model_b',
-        'variable': 'various'
-    }
-
-
-def test_join_parameters_max_count_label():
-    assert join_parameters(parameters, 2, 'label') == {
-        'model': 'model_a+model_b',
-        'variable': 'label'
-    }
-
-
-def test_copy_placeholders():
-    assert copy_placeholders({'foo': 'bar'}, {'egg': 'spam'}) == {
-        'foo': 'bar',
-        'egg': 'spam'
-    }
-
-
-def test_update_year():
-    placeholders = {'year': 2000}
-
-    update_year(placeholders, 'year', 2001, '<')
-    assert placeholders == {'year': 2000}
-
-    update_year(placeholders, 'year', 2001, '>')
-    assert placeholders == {'year': 2001}
-
-    update_year(placeholders, 'year', 2000, '<')
-    assert placeholders == {'year': 2000}
